+++ labbook_BE/app/services/ExportRest.py
@@ -3,6 +3,7 @@ import logging
 import gettext
 import csv
 import re
+import json
 
 from datetime import datetime, timedelta
 from flask import request