+++ CHANGELOG.md
@@ -8,6 +8,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - clean async process for import analysis repository
 - edit user able to upload file like CV, diploma ... as for edit staff GUI
 
+## [3.6.4] - 2025-11-03
+### Added
+- 4 new rigths for automation
+- settings for automation (dhis2, activity report and billing status)
+- template for billing status
+
+### Changed
+- redesign of the top menu
+
+### Fixed
+- big issue with reference languages
+
 ## [3.6.3] - 2025-10-23
 ### Added
 - background automation scheduler initialization in gunicorn.sh